+++ source/B2B.Transactions.CimMessageAdapter/Messages/MessageParser.cs
@@ -20,6 +20,7 @@ using System.Xml;
 using System.Xml.Schema;
 using B2B.CimMessageAdapter.Errors;
 using B2B.Transactions.IncomingMessages;
+using B2B.Transactions.IncomingMessages.RequestChangeOfSupplier;
 using B2B.Transactions.Xml.Incoming;
 
 namespace B2B.CimMessageAdapter.Messages
