+++ source/B2B.Transactions/IncomingMessages/RequestChangeOfSupplier/MarketActivityRecord.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace B2B.Transactions.IncomingMessages
+namespace B2B.Transactions.IncomingMessages.RequestChangeOfSupplier
 {
     public class MarketActivityRecord
     {
